requirements.txt
selenium
pytest
Appium-Python-Client
allure-pytest